added floats

+++ pytorch_data_loader.py
@@ -177,7 +177,7 @@ class WikiDataset(Dataset):
         #claim = sparse.vstack(claim).toarray()  # turn it into a array
         claim = self.claims_dict[d['claim']]
         claim = claim.toarray()
-        claim = torch.from_numpy(claim).cuda()
+        claim = torch.from_numpy(claim).cuda().float()
         claim_text = d['claim']
         #claim = sparse.vstack(self.encoder.tokenize_claim(utils.preprocess_article_name(d['claim']))).toarray()
 
@@ -202,7 +202,7 @@ class WikiDataset(Dataset):
                 evidence = sparse.vstack(evidence)
 
             evidence = evidence.toarray()
-            evidence = torch.from_numpy(evidence).cuda()
+            evidence = torch.from_numpy(evidence).cuda().float()
 
             evidence_text.append(processed)
             evidence_tensors.append(evidence)
@@ -238,7 +238,7 @@ class WikiDataset(Dataset):
 
             if evidence.shape[0]>0:
                 evidence = evidence.toarray()
-                evidence = torch.from_numpy(evidence).cuda()
+                evidence = torch.from_numpy(evidence).cuda().float()
                 evidence_tensors.append(evidence)
                 evidence_text.append(processed)
